refactor dashboard components imports
